backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.7
supabase==2.0.2
google-generativeai==0.3.2
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
markdownify==0.11.6
youtube-transcript-api==0.6.1
requests==2.31.0
beautifulsoup4==4.12.2
aiofiles==24.1.0
Pillow==10.1.0
pydantic==2.5.0
httpx==0.24.1